# backend/api-spec.yml
openapi: 3.0.3
info:
  title: ZombieLand API
  description: |-
    API REST pour le parc d'attractions post-apocalyptique ZombieLand.
    
    Cette API permet de gérer :
    - L'authentification et les utilisateurs (JWT)
    - Les activités et attractions du parc
    - Les catégories d'activités
    - Les réservations avec règle métier J-10
    - Les tarifs et dates d'ouverture
    - Les images des attractions
    
    **Règle métier critique** : Annulation de réservation possible uniquement si la visite est dans plus de 10 jours (sauf administrateurs).
    
    **Équipe Backend** : Jonathan Bous- Houd Alami  
  version: 1.0.0
  contact:
    name: Équipe ZombieLand
servers:
  - url: http://localhost:3001/api/v1
    description: Serveur de développement local
  - url: https://api.zombieland.com/api/v1
    description: Serveur de production
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Liste de tous les utilisateurs (ADMIN)
      description: Récupère tous les utilisateurs avec pagination et filtres
      operationId: getAllUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: email
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: Liste des utilisateurs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserDto'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Accès interdit (non ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Détail d'un utilisateur (ADMIN)
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Détail utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Accès interdit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Utilisateur non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    
    delete:
      tags:
        - Users
      summary: Supprimer un utilisateur (ADMIN)
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Utilisateur supprimé
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Accès interdit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Utilisateur non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /users/{id}/reservations:
    get:
      tags:
        - Users
      summary: Réservations d'un utilisateur (ADMIN)
      operationId: getUserReservations
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Liste des réservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationDto'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Accès interdit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Utilisateur non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

  # ===== CATEGORIES =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via /auth/login
  schemas:
...
    Role:
      type: string
      enum:
        - ADMIN
        - CLIENT
      description: Rôle utilisateur
    ReservationStatus:
      type: string
      enum:
        - PENDING
        - CONFIRMED
        - CANCELLED
      description: Statut d'une réservation
    PriceType:
      type: string
      enum:
        - ETUDIANT
        - ADULTE
        - GROUPE
        - PASS_2J
      description: Type de tarif

    # ===== RESPONSE DTOs (OUTPUT) =====
    UserDto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: user@zombieland.com
        pseudo:
          type: string
          example: JeanZ
        role:
          $ref: '#/components/schemas/Role'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ParkDateDto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        jour:
          type: string
          format: date
          example: "2025-12-25"
        is_open:
          type: boolean
          example: true
        open_hour:
          type: string
          format: time
          nullable: true
          example: "09:00:00"
          description: Heure d'ouverture du parc (format HH:mm:ss)
        close_hour:
          type: string
          format: time
          nullable: true
          example: "22:00:00"
          description: Heure de fermeture du parc (format HH:mm:ss)
        notes:
          type: string
          nullable: true
          example: Horaires étendus pour les fêtes
        created_at:
          type: string
          format: date-time
    PriceDto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        label:
          type: string
          example: Tarif Étudiant
        type:
          $ref: '#/components/schemas/PriceType'
        amount:
          type: number
          format: decimal
          example: 29.99
        duration_days:
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ReservationDto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        reservation_number:
          type: string
          example: ZL-1732108800-A7F3B
        user_id:
          type: integer
          example: 5
        date_id:
          type: integer
          example: 12
        price_id:
          type: integer
          example: 2
        tickets_count:
          type: integer
          example: 3
        total_amount:
          type: number
          format: decimal
          example: 89.97
        status:
          $ref: '#/components/schemas/ReservationStatus'
        user:
          $ref: '#/components/schemas/UserDto'
        date:
          $ref: '#/components/schemas/ParkDateDto'
        price:
          $ref: '#/components/schemas/PriceDto'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ===== REQUEST DTOs (INPUT) =====
...
    ErrorDto:
      type: object
      required:
        - statusCode
        - message
        - error
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Validation failed
        error:
          type: string
          example: Bad Request
        timestamp:
          type: string
          format: date-time